Separate out new and edit branch errors acceptance tests
- Move the creation of a new branch page to its' own test file.
- Check for correct paths for '/new' and '/edit' before and after saving
- Check correctly selected operator and answer are chosen by default
- Move shared steps to BranchingSteps module
