+++ acceptance/spec_helper.rb
@@ -31,6 +31,7 @@ RSpec.configure do |config|
 
   config.include TestHelpers
   config.include CommonSteps
+  config.include BranchingSteps
   config.include DataContentId
   config.include MultipleQuestionsPageHelper
 end
